split item and model
